edit category migration and seeder

+++ database/migrations/2022_12_09_182458_create_lots_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->float('price');
+            $table->integer('price');
             $table->timestamps();
         });
     }
